// front-end/petproposal-front/src/app/model/user-info.ts
export class UserInfo {
    userSpace: string;
    userTime: string;
    bearableNoise: string;
    alergicToHair: boolean;
    alergicToFeather: boolean;
    moneyForSpendingOnAnimal: string;
    animalThatLikesPetting: boolean;
    absentFromHome: string;
    outdoorActivities: boolean;
    teachingTricks: boolean;

}